// src/shared/generated/ChromiumPopupLifecycleActionRecord.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ChromiumPopupCloseReason } from "./ChromiumPopupCloseReason";
import type { ChromiumPopupNativeHostReceiptRecord } from "./ChromiumPopupNativeHostReceiptRecord";

export type ChromiumPopupLifecycleActionRecord = { "type": "nativeReady", host: ChromiumPopupNativeHostReceiptRecord, } | { "type": "pageReady", finalUrl: string, } | { "type": "closeRequested", reason: ChromiumPopupCloseReason, } | { "type": "nativeClosed" } | { "type": "cancelled", failureCode: string, } | { "type": "failed", failureCode: string, nativeStateUnknown: boolean, };